+++ packages/create-react-native-library/templates/native-common-example/example/babel.config.js
@@ -1,17 +1,12 @@
 const path = require('path');
-const pak = require('../package.json');
+const { getConfig } = require('react-native-builder-bob/babel-config');
+const pkg = require('../package.json');
 
-module.exports = {
-  presets: ['module:@react-native/babel-preset'],
-  plugins: [
-    [
-      'module-resolver',
-      {
-        extensions: ['.tsx', '.ts', '.js', '.json'],
-        alias: {
-          [pak.name]: path.join(__dirname, '..', pak.source),
-        },
-      },
-    ],
-  ],
-};
+const root = path.resolve(__dirname, '..');
+
+module.exports = getConfig(
+  {
+    presets: ['module:@react-native/babel-preset'],
+  },
+  { root, pkg }
+);