+++ packages/react-native-builder-bob/src/index.ts
@@ -3,7 +3,7 @@ import fs from 'fs-extra';
 import kleur from 'kleur';
 import dedent from 'dedent';
 import yargs from 'yargs';
-import { cosmiconfigSync } from 'cosmiconfig';
+import { cosmiconfig } from 'cosmiconfig';
 import isGitDirty from 'is-git-dirty';
 import prompts, { type PromptObject } from './utils/prompts';
 import * as logger from './utils/logger';
@@ -27,8 +27,14 @@ const args: Record<ArgName, yargs.Options> = {
 const { name, version } = require('../package.json');
 
 const root = process.cwd();
-const explorer = cosmiconfigSync(name, {
-  searchPlaces: ['package.json', `bob.config.js`, 'bob.config.cjs'],
+const explorer = cosmiconfig(name, {
+  stopDir: root,
+  searchPlaces: [
+    'package.json',
+    'bob.config.mjs',
+    'bob.config.cjs',
+    'bob.config.js',
+  ],
 });
 
 const FLOW_PRGAMA_REGEX = /\*?\s*@(flow)\b/m;
@@ -57,7 +63,7 @@ yargs
     }
 
     const pkg = JSON.parse(await fs.readFile(pak, 'utf-8'));
-    const result = explorer.search();
+    const result = await explorer.search();
 
     if (result?.config && pkg.devDependencies && name in pkg.devDependencies) {
       const { shouldContinue } = await prompts({
@@ -170,23 +176,38 @@ yargs
       esm = true;
 
       if (targets.includes('commonjs')) {
-        entries.main = `./${path.join(output, 'commonjs', 'index.cjs')}`;
+        entries.main = `./${path.join(output, 'commonjs', 'index.js')}`;
       }
 
-      entries.module = `./${path.join(output, 'module', 'index.mjs')}`;
+      entries.module = `./${path.join(output, 'module', 'index.js')}`;
     } else if (targets.includes('commonjs')) {
       entries.main = `./${path.join(output, 'commonjs', 'index.js')}`;
     } else {
       entries.main = entries.source;
     }
 
+    const types: {
+      [key in 'require' | 'import']?: string;
+    } = {};
+
     if (targets.includes('typescript')) {
-      entries.types = `./${path.join(
+      types.require = `./${path.join(
         output,
         'typescript',
+        'commonjs',
         source,
         'index.d.ts'
       )}`;
+
+      types.import = `./${path.join(
+        output,
+        'typescript',
+        'module',
+        source,
+        'index.d.ts'
+      )}`;
+
+      entries.types = types.require;
 
       if (!(await fs.pathExists(path.join(root, 'tsconfig.json')))) {
         const { tsconfig } = await prompts({
@@ -263,9 +284,14 @@ yargs
 
       const exports = {
         '.': {
-          ...(entries.types ? { types: entries.types } : null),
-          ...(entries.module ? { import: entries.module } : null),
-          ...(entries.main ? { require: entries.main } : null),
+          import: {
+            ...(types.import ? { types: types.import } : null),
+            ...(entries.module ? { default: entries.module } : null),
+          },
+          require: {
+            ...(types.require ? { types: types.require } : null),
+            ...(entries.main ? { default: entries.main } : null),
+          },
         },
       };
 
@@ -323,25 +349,23 @@ yargs
       pkg.scripts.prepare = prepare;
     }
 
-    if (
-      pkg.files &&
-      JSON.stringify(pkg.files.slice().sort()) !==
-        JSON.stringify(files.slice().sort())
-    ) {
-      const { update } = await prompts({
-        type: 'confirm',
-        name: 'update',
-        message: `Your package.json already has a 'files' field.\n  Do you want to update it?`,
-        initial: true,
-      });
+    if (pkg.files) {
+      const pkgFiles = pkg.files;
 
-      if (update) {
-        pkg.files = [
-          ...files,
-          ...pkg.files.filter(
-            (file: string) => !files.includes(file.replace(/\/$/g, ''))
-          ),
-        ];
+      if (files?.some((file) => !pkgFiles.includes(file))) {
+        const { update } = await prompts({
+          type: 'confirm',
+          name: 'update',
+          message: `Your package.json already has a 'files' field.\n  Do you want to update it?`,
+          initial: true,
+        });
+
+        if (update) {
+          pkg.files = [
+            ...files,
+            ...pkg.files.filter((file: string) => !files.includes(file)),
+          ];
+        }
       }
     } else {
       pkg.files = files;
@@ -355,7 +379,7 @@ yargs
           return [t, { copyFlow: true }];
         }
 
-        if (t === 'commonjs' || t === 'module') {
+        if (t === 'commonjs' || t === 'module' || t === 'typescript') {
           return [t, { esm }];
         }
 
@@ -423,8 +447,8 @@ yargs
     `)
     );
   })
-  .command('build', 'build files for publishing', args, async (argv) => {
-    const result = explorer.search();
+  .command('build', 'build files for publishing', {}, async (argv) => {
+    const result = await explorer.search();
 
     if (!result?.config) {
       logger.exit(